:tada: chore: refactor modals
